refactor: move MDA data types to dedicated module
Move FeaturedItem, MinistryService, ContactItem, Minister, and
AssociatedDepartmentGroup out of components/ministry/ministry-page.tsx
into src/data/mda-types.ts so data files no longer import shapes from
presentation layer. Drop duplicate AssociatedDepartmentGroup declaration
in data/ministries.ts.
